# CUDA: Faster prompt processing for several quantization types 